Adjust userland prototypes as well.

+++ sys/sys/tls.h
@@ -3,7 +3,7 @@
  *
  *	Implements the architecture independant TLS info structure.
  *
- * $DragonFly: src/sys/sys/tls.h,v 1.2 2005/02/22 02:17:56 dillon Exp $
+ * $DragonFly: src/sys/sys/tls.h,v 1.3 2005/03/21 17:09:52 joerg Exp $
  */
 
 #ifndef _SYS_TLS_H_
@@ -15,8 +15,8 @@ struct tls_info {
 };
 
 #ifndef _KERNEL
-int sys_set_tls_area(int which, struct tls_info *info, int infosize);
-int sys_get_tls_area(int which, struct tls_info *info, int infosize);
+int sys_set_tls_area(int which, struct tls_info *info, size_t infosize);
+int sys_get_tls_area(int which, struct tls_info *info, size_t infosize);
 #endif
 
 #endif
